requirements.txt
allure-pytest==2.9.45
chromedriver-binary-auto==0.1.1
funcy==1.17
pytest==7.1.1
python-dotenv==0.19.1
requests==2.27.1
selenium==4.1.3
